# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API APP",
        "description": "API Documentation",
        "contact": {
            "email": "chiranad@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Courses",
                "requestBody": {
                    "description": "Courses list",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "success": {
                                        "type": "boolean",
                                        "example": "true"
                                    },
                                    "payload": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "int",
                                                    "example": ""
                                                },
                                                "code": {
                                                    "type": "string",
                                                    "example": ""
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": ""
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": ""
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": ""
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": ""
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "example": [
                                            {
                                                "id": 1,
                                                "code": "059-209-6781",
                                                "name": "IT",
                                                "description": "IT",
                                                "created_at": "2022-01-13T19:00:17.000000Z",
                                                "updated_at": "2022-01-13T19:00:17.000000Z"
                                            },
                                            {
                                                "id": 1,
                                                "code": "059-209-6781",
                                                "name": "English",
                                                "description": "English",
                                                "created_at": "2022-01-13T19:00:17.000000Z",
                                                "updated_at": "2022-01-13T19:00:17.000000Z"
                                            }
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/api/student-courses": {
            "get": {
                "tags": [
                    "Student Courses"
                ],
                "summary": "Student Courses",
                "requestBody": {
                    "description": "Student Courses list",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "success": {
                                        "type": "boolean",
                                        "example": "true"
                                    },
                                    "payload": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "int",
                                                    "example": ""
                                                },
                                                "code": {
                                                    "type": "string",
                                                    "example": ""
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": ""
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": ""
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": ""
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": ""
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "example": [
                                            {
                                                "id": 1,
                                                "code": "059-209-6781",
                                                "name": "IT",
                                                "description": "IT",
                                                "created_at": "2022-01-13T19:00:17.000000Z",
                                                "updated_at": "2022-01-13T19:00:17.000000Z"
                                            },
                                            {
                                                "id": 1,
                                                "code": "059-209-6781",
                                                "name": "English",
                                                "description": "English",
                                                "created_at": "2022-01-13T19:00:17.000000Z",
                                                "updated_at": "2022-01-13T19:00:17.000000Z"
                                            }
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/courses/enroll": {
            "post": {
                "tags": [
                    "Enrollement"
                ],
                "summary": "Enroll Courses",
                "operationId": "enroll",
                "parameters": [
                    {
                        "name": "course_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "successfully register"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "payload": {
                                            "properties": {
                                                "id": {
                                                    "type": "int",
                                                    "example": "1"
                                                },
                                                "code": {
                                                    "type": "string",
                                                    "example": "327-829-8548"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "IT"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Rem quos esse est voluptates rerum"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2022-01-13T18:16:24.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2022-01-13T18:16:24.000000Z"
                                                }
                                            },
                                            "type": "object",
                                            "format": "query"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}